// src/@types/qraphql.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Todo = {
  __typename?: 'Todo';
  id: Scalars['String'];
  task: Scalars['String'];
  isComplete: Scalars['Boolean'];
};

export type TodoInput = {
  id: Scalars['String'];
  task: Scalars['String'];
  isComplete: Scalars['Boolean'];
};

export type ResponseMessage = {
  __typename?: 'ResponseMessage';
  success: Scalars['Boolean'];
};

export type Id = {
  __typename?: 'Id';
  id: Scalars['String'];
};

export type TodoDeleteResponse = {
  __typename?: 'TodoDeleteResponse';
  success: Scalars['Boolean'];
  data: Id;
};

export type TodoUpdateResponse = {
  __typename?: 'TodoUpdateResponse';
  success: Scalars['Boolean'];
  data: Todo;
};

export type Query = {
  __typename?: 'Query';
  todos: Array<Maybe<Todo>>;
};

export type TodoUpdateAllResponse = {
  __typename?: 'TodoUpdateAllResponse';
  success: Scalars['Boolean'];
  data?: Maybe<Query>;
};

export type Mutation = {
  __typename?: 'Mutation';
  createTodo: TodoUpdateResponse;
  deleteTodo: TodoDeleteResponse;
  deleteAllTodos: ResponseMessage;
  editTodo: TodoUpdateResponse;
  toggleTodo: TodoUpdateResponse;
  updateAllTodos: TodoUpdateAllResponse;
};


export type MutationCreateTodoArgs = {
  task: Scalars['String'];
};


export type MutationDeleteTodoArgs = {
  id: Scalars['String'];
};


export type MutationEditTodoArgs = {
  id: Scalars['String'];
  task: Scalars['String'];
};


export type MutationToggleTodoArgs = {
  id: Scalars['String'];
  isComplete: Scalars['Boolean'];
};


export type MutationUpdateAllTodosArgs = {
  todos: Array<Maybe<TodoInput>>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
  selectionSet: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type StitchingResolver<TResult, TParent, TContext, TArgs> = LegacyStitchingResolver<TResult, TParent, TContext, TArgs> | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = {
  Todo: ResolverTypeWrapper<Todo>;
  String: ResolverTypeWrapper<Scalars['String']>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  TodoInput: TodoInput;
  ResponseMessage: ResolverTypeWrapper<ResponseMessage>;
  Id: ResolverTypeWrapper<Id>;
  TodoDeleteResponse: ResolverTypeWrapper<TodoDeleteResponse>;
  TodoUpdateResponse: ResolverTypeWrapper<TodoUpdateResponse>;
  Query: ResolverTypeWrapper<{}>;
  TodoUpdateAllResponse: ResolverTypeWrapper<TodoUpdateAllResponse>;
  Mutation: ResolverTypeWrapper<{}>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Todo: Todo;
  String: Scalars['String'];
  Boolean: Scalars['Boolean'];
  TodoInput: TodoInput;
  ResponseMessage: ResponseMessage;
  Id: Id;
  TodoDeleteResponse: TodoDeleteResponse;
  TodoUpdateResponse: TodoUpdateResponse;
  Query: {};
  TodoUpdateAllResponse: TodoUpdateAllResponse;
  Mutation: {};
};
// ...
export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  createTodo?: Resolver<ResolversTypes['TodoUpdateResponse'], ParentType, ContextType, RequireFields<MutationCreateTodoArgs, 'task'>>;
  deleteTodo?: Resolver<ResolversTypes['TodoDeleteResponse'], ParentType, ContextType, RequireFields<MutationDeleteTodoArgs, 'id'>>;
  deleteAllTodos?: Resolver<ResolversTypes['ResponseMessage'], ParentType, ContextType>;
  editTodo?: Resolver<ResolversTypes['TodoUpdateResponse'], ParentType, ContextType, RequireFields<MutationEditTodoArgs, 'id' | 'task'>>;
  toggleTodo?: Resolver<ResolversTypes['TodoUpdateResponse'], ParentType, ContextType, RequireFields<MutationToggleTodoArgs, 'id' | 'isComplete'>>;
  updateAllTodos?: Resolver<ResolversTypes['TodoUpdateAllResponse'], ParentType, ContextType, RequireFields<MutationUpdateAllTodosArgs, 'todos'>>;
};